+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [5.0.10](https://github.com/rizen/pixi-glyphs/compare/v5.0.9...v5.0.10) (2025-12-30)
+
 ### [5.0.9](https://github.com/rizen/pixi-glyphs/compare/v5.0.8...v5.0.9) (2025-12-15)
 
 - Updated the Spacing demo to be able to change the text.